added method to add and update supplier stock

+++ Backend/PharmacyProject/src/main/java/com/team11/PharmacyProject/dto/supplier/SupplierStockItemDTO.java
@@ -3,12 +3,12 @@ package com.team11.PharmacyProject.dto.supplier;
 public class SupplierStockItemDTO {
     private long medicineId;
     private String medicineName;
-    private double amount;
+    private int amount;
 
     public SupplierStockItemDTO() {
     }
 
-    public SupplierStockItemDTO(long medicineId, String medicineName, double totalPrice) {
+    public SupplierStockItemDTO(long medicineId, String medicineName, int totalPrice) {
         this.medicineId = medicineId;
         this.medicineName = medicineName;
         this.amount = totalPrice;
@@ -30,11 +30,11 @@ public class SupplierStockItemDTO {
         this.medicineName = medicineName;
     }
 
-    public double getAmount() {
+    public int getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
 }
